Support tablet mode in material & samsung & fix widths across screens when in tablet mode
