+++ lib/layouts/image_viewer/image_viewer.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 import 'dart:typed_data';
 
+import 'package:bluebubbles/helpers/navigator.dart';
 import 'package:bluebubbles/helpers/utils.dart';
 import 'package:get/get.dart';
 import 'package:bluebubbles/helpers/attachment_helper.dart';
@@ -75,7 +76,7 @@ class _ImageViewerState extends State<ImageViewer> with AutomaticKeepAliveClient
       duration: Duration(milliseconds: 125),
       child: Container(
           height: 150.0,
-          width: context.width,
+          width: CustomNavigator.width(context),
           color: Colors.black.withOpacity(0.65),
           child: Row(mainAxisAlignment: MainAxisAlignment.spaceBetween, children: [
             Padding(
@@ -128,7 +129,7 @@ class _ImageViewerState extends State<ImageViewer> with AutomaticKeepAliveClient
                           ),
                           backgroundColor: Theme.of(context).accentColor,
                           content: SizedBox(
-                            width: context.width * 3 / 5,
+                            width: CustomNavigator.width(context) * 3 / 5,
                             height: context.height * 1 / 4,
                             child: Container(
                               padding: EdgeInsets.all(10.0),
