+++ lib/layouts/search/search_view.dart
@@ -148,7 +148,7 @@ class SearchViewState extends State<SearchView> with TickerProviderStateMixin {
         // extendBodyBehindAppBar: true,
         backgroundColor: Theme.of(context).backgroundColor,
         appBar: PreferredSize(
-          preferredSize: Size(context.width, 80),
+          preferredSize: Size(CustomNavigator.width(context), 80),
           child: ClipRRect(
             child: BackdropFilter(
               child: AppBar(
