+++ lib/layouts/settings/custom_avatar_panel.dart
@@ -50,7 +50,7 @@ class CustomAvatarPanel extends GetView<CustomAvatarPanelController> {
       child: Scaffold(
         backgroundColor: Theme.of(context).backgroundColor,
         appBar: PreferredSize(
-          preferredSize: Size(context.width, 80),
+          preferredSize: Size(CustomNavigator.width(context), 80),
           child: ClipRRect(
             child: BackdropFilter(
               child: AppBar(
