+++ lib/layouts/settings/theme_panel.dart
@@ -93,7 +93,7 @@ class ThemePanel extends GetView<ThemePanelController> {
         child: Scaffold(
           backgroundColor: SettingsManager().settings.skin.value != Skins.iOS ? tileColor : headerColor,
           appBar: PreferredSize(
-            preferredSize: Size(context.width, 80),
+            preferredSize: Size(CustomNavigator.width(context), 80),
             child: ClipRRect(
               child: BackdropFilter(
                 child: AppBar(
