+++ lib/layouts/widgets/avatar_crop.dart
@@ -3,6 +3,7 @@ import 'dart:typed_data';
 import 'dart:ui';
 
 import 'package:bluebubbles/blocs/chat_bloc.dart';
+import 'package:bluebubbles/helpers/navigator.dart';
 import 'package:bluebubbles/helpers/ui_helpers.dart';
 import 'package:bluebubbles/helpers/utils.dart';
 import 'package:bluebubbles/managers/method_channel_interface.dart';
@@ -66,7 +67,7 @@ class _AvatarCropState extends State<AvatarCrop> {
       child: Scaffold(
           backgroundColor: Theme.of(context).backgroundColor,
           appBar: PreferredSize(
-            preferredSize: Size(context.width, 80),
+            preferredSize: Size(CustomNavigator.width(context), 80),
             child: ClipRRect(
               child: BackdropFilter(
                 child: AppBar(
