+++ lib/layouts/widgets/contact_avatar_widget.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 
 import 'package:bluebubbles/helpers/constants.dart';
 import 'package:bluebubbles/helpers/hex_color.dart';
+import 'package:bluebubbles/helpers/navigator.dart';
 import 'package:bluebubbles/helpers/utils.dart';
 import 'package:bluebubbles/managers/contact_manager.dart';
 import 'package:bluebubbles/managers/settings_manager.dart';
@@ -162,7 +163,7 @@ class _ContactAvatarWidgetState extends State<ContactAvatarWidget> with Automati
       context,
       widget.handle?.color != null ? HexColor(widget.handle!.color!) : toColorGradient(widget.handle!.address)[0],
       title: Container(
-          width: context.width - 112,
+          width: CustomNavigator.width(context) - 112,
           child: Row(
               mainAxisAlignment: MainAxisAlignment.spaceBetween,
               children: [
@@ -200,7 +201,7 @@ class _ContactAvatarWidgetState extends State<ContactAvatarWidget> with Automati
         dialogActionButtons: true,
       ),
       constraints: BoxConstraints(
-          minHeight: 480, minWidth: context.width - 70, maxWidth: context.width - 70),
+          minHeight: 480, minWidth: CustomNavigator.width(context) - 70, maxWidth: CustomNavigator.width(context) - 70),
     );
 
     if (didReset) return;
