+++ lib/layouts/widgets/message_widget/message_content/media_players/audio_player_widget.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 
 import 'package:bluebubbles/helpers/constants.dart';
+import 'package:bluebubbles/helpers/navigator.dart';
 import 'package:bluebubbles/managers/settings_manager.dart';
 import 'package:chewie_audio/chewie_audio.dart';
 import 'package:get/get.dart';
@@ -91,7 +92,7 @@ class _AudioPlayerWigetState extends State<AudioPlayerWiget> with AutomaticKeepA
 
   @override
   Widget build(BuildContext context) {
-    double maxWidth = widget.width ?? context.width - 20;
+    double maxWidth = widget.width ?? CustomNavigator.width(context) - 20;
     if (!(ModalRoute.of(context)?.isCurrent ?? false)) {
       controller.pause();
     }
