+++ lib/layouts/widgets/message_widget/message_content/message_attachment.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 
+import 'package:bluebubbles/helpers/navigator.dart';
 import 'package:bluebubbles/helpers/ui_helpers.dart';
 import 'package:get/get.dart';
 import 'package:bluebubbles/helpers/attachment_downloader.dart';
@@ -69,7 +70,7 @@ class MessageAttachmentState extends State<MessageAttachment> with AutomaticKeep
       borderRadius: BorderRadius.circular(20),
       child: Container(
         constraints: BoxConstraints(
-          maxWidth: context.width * 0.5,
+          maxWidth: CustomNavigator.width(context) * 0.5,
           maxHeight: context.height * 0.6,
         ),
         child: _buildAttachmentWidget(),
